qrcode&captcha/requirements.txt
qrcode
pillow
pyzbar
cv2